service MS_VARS_FREEMEMORY replaced with the direct call of mir_free
git-svn-id: http://svn.miranda-ng.org/main/trunk@3669 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Variables/src/parse_regexp.cpp
@@ -46,7 +46,7 @@ static TCHAR *parseRegExpCheck(ARGUMENTSINFO *ai) {
 		return NULL;
 	}
 	pcre_extra *extra = pcre_study(ppat, 0, &err);
-	int nmat = pcre_exec(ppat, extra, arg2, strlen(arg2), 0, 0, offsets, 99);
+	int nmat = pcre_exec(ppat, extra, arg2, (int)strlen(arg2), 0, 0, offsets, 99);
 	mir_free(arg1);
 	mir_free(arg2);
 	if (nmat > 0) {
@@ -96,7 +96,7 @@ static TCHAR *parseRegExpSubstr(ARGUMENTSINFO *ai) {
 		return NULL;
 	}
 	pcre_extra *extra = pcre_study(ppat, 0, &err);
-	int nmat = pcre_exec(ppat, extra, arg2, strlen(arg2), 0, 0, offsets, 99);
+	int nmat = pcre_exec(ppat, extra, arg2, (int)strlen(arg2), 0, 0, offsets, 99);
 	if (nmat >= 0) {
 		ai->flags &= ~AIF_FALSE;
 	}
